+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.9.0] / 2025-02-07
+## [1.9.0] / 2025-02-07 - 2025-02-19
 ### Features
 - Update `ricaun.NUnit` support `TestFixture` - Version 1.5.0.
 - Update `ricaun.Revit.UI` version 0.8.0.